Debut c++
git-svn-id: https://scm.gforge.inria.fr/svn/mitomer/trunk@18 03fad38e-d51a-4bec-b665-4e7f05cec1f8

+++ count_kmer/src/count_lib.h
@@ -2,7 +2,7 @@
 #define COUNT_LIB_H_
 
 
-unsigned long long puissance4(unsigned long long x);
+int puissance4(int x);
 
 /**
  * Vérifie qu'un pointeur est non null
@@ -16,14 +16,14 @@ void my_error(void *ptr, const char *erreur);
  * Retourne val kmer
  * 
  * */
- char *indice_to_kmer(unsigned long long indice, long kmer);  
+ char *indice_to_kmer(int indice, long kmer);  
 
 /**
  *
  *
  *
  **/
- void init(unsigned long long *tab_comptage,int taille_kmer);
+ void init(int *tab_comptage,int taille_kmer);
 /**
  *
  *
@@ -31,19 +31,19 @@ void my_error(void *ptr, const char *erreur);
  *
  **/
 
-void concat_one(unsigned long long *tab_comptage,int taille_kmer,char *dest, int i);
+void concat_one(int *tab_comptage,int taille_kmer,char *dest, int i);
 
-void concat_two(unsigned long long *tab_comptage,char *dest,int j);
+void concat_two(int *tab_comptage,char *dest,int j);
 
-void affiche(unsigned long long *tab_comptage,int taille_kmer);
+void affiche(int *tab_comptage,int taille_kmer);
 
-void add_one(const char *seq,int i,const size_t seq_taille,int pattern_taille,unsigned long long *resultat,int *tab_bool,int kmer_taille);
+void add_one(const char *seq,int i,const size_t seq_taille,int pattern_taille,int *resultat,int *tab_bool,int kmer_taille);
 
-void count_seq(const char *seq, const size_t seq_taille,int *tab_bool,unsigned long long *resultat,int pattern_taille) ;
+void count_seq(const char *seq, const size_t seq_taille,int *tab_bool,int *resultat,int pattern_taille) ;
 
-void count(const char *seq,const size_t seq_taille,int *tab_bool,int *taille_kmer,unsigned long long ** resultat,int taille_sous_sequence);
+void count(const char *seq,const size_t seq_taille,int *tab_bool,int *taille_kmer,int ** resultat,int taille_sous_sequence);
 
-// void count(const char *seq,const size_t seq_taille,const char *kmer_string,int kmer_taille,int pattern_taille,unsigned long long *tab_comptage,int taille_sous_sequence);
+// void count(const char *seq,const size_t seq_taille,const char *kmer_string,int kmer_taille,int pattern_taille,int *tab_comptage,int taille_sous_sequence);
