+++ count_kmer/src/count_print.c
@@ -19,7 +19,7 @@ void imprime_kmer(FILE *out,int taille_kmer)
 
 	int i;
 
-	unsigned long long taille =  puissance4(taille_kmer) ;
+	int taille =  puissance4(taille_kmer) ;
 
 	for (i = 0; i < taille; i++)
 	{
@@ -29,11 +29,11 @@ void imprime_kmer(FILE *out,int taille_kmer)
 	
 }
 
-void imprime_csv(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,unsigned long long ***resultat)
+void imprime_csv(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,int ***resultat)
 {
 	int i,j,k;
 	FILE *out = NULL;
-	unsigned long long nb_possibilite;
+	int nb_possibilite;
 
 
 	out = fopen(path, "a");
@@ -52,11 +52,11 @@ void imprime_csv(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_s
 			for( k= 0;k<nb_kmer;k++)
 			{
 				nb_possibilite = puissance4(taille_kmer[k]);
-				// printf("nb_possibilite = %llu\n",nb_possibilite);
+				// printf("nb_possibilite = %d\n",nb_possibilite);
 
 				for (j = 0; j < nb_possibilite; j++)
 				{
-					fprintf(out,"%llu;",resultat[k][i][j]);
+					fprintf(out,"%d;",resultat[k][i][j]);
 				}
 				
 			}
@@ -75,9 +75,9 @@ void imprime_csv(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_s
 
 void imprime_entete_weka(FILE *fi,int taille_kmer)
 {
-	unsigned long long i;
+	int i;
 
-	unsigned long long taille =  puissance4(taille_kmer) ;
+	int taille =  puissance4(taille_kmer) ;
 
 	for (i = 0; i < taille; i++)
 	{
@@ -85,11 +85,11 @@ void imprime_entete_weka(FILE *fi,int taille_kmer)
 	}
 }
 
-void imprime_weka(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,unsigned long long ***resultat,const char *taxid,int flag_en_tete,char *les_accessions)
+void imprime_weka(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,int ***resultat,const char *taxid,int flag_en_tete,char *les_accessions)
 {
-	unsigned long long i,j,k;
+	int i,j,k;
 	FILE *out = NULL;
-	unsigned long long nb_possibilite;
+	int nb_possibilite;
 
 	// printf("JECRIS\n");
 	out = fopen(path, "a");
@@ -114,12 +114,12 @@ void imprime_weka(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_
 			for( k= 0;k<nb_kmer;k++)
 			{
 				nb_possibilite = puissance4(taille_kmer[k]);
-				// printf("nb_possibilite = %llu\n",nb_possibilite);
+				// printf("nb_possibilite = %d\n",nb_possibilite);
 
 				for (j = 0; j < nb_possibilite; j++)
 				{
-					fprintf(out,"%llu,",resultat[k][i][j]);
-					// fprintf(stdout,"%llu,",resultat[k][i][j]);
+					fprintf(out,"%d,",resultat[k][i][j]);
+					// fprintf(stdout,"%d,",resultat[k][i][j]);
 
 				}
