+++ count_kmer/src/count_print.h
@@ -2,16 +2,16 @@
 #define COUNT_PRINT_H_
 
 
-void imprime(unsigned long long **tab_comptage,int *taille_kmer);
+void imprime(int **tab_comptage,int *taille_kmer);
 
-void imprime_weka(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,unsigned long long ***resultat,const char *taxid,int flag,char *les_accessions);
+void imprime_weka(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,int ***resultat,const char *taxid,int flag,char *les_accessions);
 
-void imprime_csv(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,unsigned long long ***resultat);
+void imprime_csv(const char *path,int nb_kmer,int *taille_kmer,int nombre_sous_sequence,int ***resultat);
 
 void imprime_kmer(FILE *out,int taille_kmer);
 
 void imprime_entete_weka(FILE *fi,int taille_kmer);
 
-char *indice_to_kmer(unsigned long long indice, long kmer);
+char *indice_to_kmer(int indice, long kmer);
 
 #endif
